refactor(patina): move no-top-level-ref tests into their own file
