+++ crates/vize_patina/src/linter/tests.rs
@@ -5,4 +5,5 @@ use vize_carton::{Allocator, ToCompactString};
 mod basic;
 mod directives;
 mod jsx;
+mod no_top_level_ref;
 mod sfc;
